+++ lib/opte-test-utils/src/icmp.rs
@@ -7,8 +7,8 @@
 //! Routines for ICMP testing.
 
 use opte::api::*;
+use opte::ddi::mblk::MsgBlk;
 use opte::engine::ether::Ethernet;
-use opte::engine::ingot_packet::MsgBlk;
 use opte::engine::ip::v4::Ipv4;
 use opte::engine::ip::v6::Ipv6;
 use opte::engine::ip::L3;
