+++ lib/opte/src/ddi/mblk.rs
@@ -3,3 +3,831 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 // Copyright 2024 Oxide Computer Company
+
+use crate::engine::ingot_packet::QueryLen;
+use crate::engine::packet::allocb;
+#[cfg(any(feature = "std", test))]
+use crate::engine::packet::mock_freemsg;
+use crate::engine::packet::SegAdjustError;
+use crate::engine::packet::WrapError;
+use crate::engine::packet::WriteError;
+use alloc::vec::Vec;
+use core::marker::PhantomData;
+use core::mem::ManuallyDrop;
+use core::mem::MaybeUninit;
+use core::ops::Deref;
+use core::ops::DerefMut;
+use core::ptr;
+use core::ptr::NonNull;
+use core::slice;
+#[cfg(all(not(feature = "std"), not(test)))]
+use illumos_sys_hdrs as ddi;
+use illumos_sys_hdrs::mblk_t;
+use illumos_sys_hdrs::uintptr_t;
+use ingot::types::Emit;
+use ingot::types::EmitDoesNotRelyOnBufContents;
+use ingot::types::ParseError as IngotParseErr;
+use ingot::types::Read;
+
+/// The head and tail of an mblk_t list.
+struct MsgBlkChainInner {
+    head: NonNull<mblk_t>,
+    tail: NonNull<mblk_t>,
+}
+
+/// A chain of illumos MsgBlk/`mblk_t` buffers.
+///
+/// Network packets are provided by illumos as a linked list of linked lists,
+/// using the `b_next` and `b_prev` fields.
+///
+/// See the documentation for [`super::ingot_packet::Packet`] and/or [`MsgBlk`] for full context.
+// TODO: We might retool this type now that MsgBlk does not decompose
+// each mblk_t into individual segments (i.e., packets could be allocated
+// a lifetime via PhantomData based on whether we want to remove them from the chain or modify in place).
+// Today's code is all equivalent to always using 'static, because
+// we remove and re-add the mblks to work on them.
+// We might want also want to return either a chain/mblk_t in an enum, but
+// practically XDE will always assume it has a chain from MAC.
+pub struct MsgBlkChain {
+    inner: Option<MsgBlkChainInner>,
+}
+
+impl MsgBlkChain {
+    /// Create an empty packet chain.
+    pub fn empty() -> Self {
+        Self { inner: None }
+    }
+
+    /// Convert an mblk_t packet chain into a safe source of `MsgBlk`s.
+    ///
+    /// # Safety
+    /// The `mp` pointer must point to an `mblk_t` allocated by
+    /// `allocb(9F)` or provided by some kernel API which itself used
+    /// one of the DDI/DKI APIs to allocate it.
+    /// Packets must form a valid linked list (no loops).
+    /// The original mblk_t pointer must not be used again.
+    pub unsafe fn new(mp: *mut mblk_t) -> Result<Self, WrapError> {
+        let head = NonNull::new(mp).ok_or(WrapError::NullPtr)?;
+
+        // Walk the chain to find the tail, and support faster append.
+        let mut tail = head;
+        while let Some(next_ptr) = NonNull::new((*tail.as_ptr()).b_next) {
+            tail = next_ptr;
+        }
+
+        Ok(Self { inner: Some(MsgBlkChainInner { head, tail }) })
+    }
+
+    /// Removes the next packet from the top of the chain and returns
+    /// it, taking ownership.
+    pub fn pop_front(&mut self) -> Option<MsgBlk> {
+        if let Some(ref mut list) = &mut self.inner {
+            unsafe {
+                let curr_b = list.head;
+                let curr = curr_b.as_ptr();
+                let next = NonNull::new((*curr).b_next);
+
+                // Break the forward link on the packet we have access to,
+                // and the backward link on the next element if possible.
+                if let Some(next) = next {
+                    (*next.as_ptr()).b_prev = ptr::null_mut();
+                }
+                (*curr).b_next = ptr::null_mut();
+
+                // Update the current head. If the next element is null,
+                // we're now empty.
+                if let Some(next) = next {
+                    list.head = next;
+                } else {
+                    self.inner = None;
+                }
+
+                Some(MsgBlk { inner: curr_b })
+            }
+        } else {
+            None
+        }
+    }
+
+    /// Adds an owned `MsgBlk` to the end of this chain.
+    ///
+    /// Internally, this unwraps the `MsgBlk` back into an mblk_t,
+    /// before placing it at the tail.
+    pub fn append(&mut self, packet: MsgBlk) {
+        // Unwrap safety: a valid Packet implies a non-null mblk_t.
+        // Jamming `NonNull` into PacketSeg/Packet might take some
+        // work just to avoid this unwrap.
+        let pkt = packet.unwrap_mblk();
+
+        // We're guaranteeing today that a 'static Packet has
+        // no neighbours and is not part of a chain.
+        // This simplifies tail updates in both cases (no chain walk).
+        unsafe {
+            assert!((*pkt.as_ptr()).b_prev.is_null());
+            assert!((*pkt.as_ptr()).b_next.is_null());
+        }
+
+        if let Some(ref mut list) = &mut self.inner {
+            let pkt_p = pkt.as_ptr();
+            let tail_p = list.tail.as_ptr();
+            unsafe {
+                (*tail_p).b_next = pkt_p;
+                (*pkt_p).b_prev = tail_p;
+                // pkt_p->b_next is already null.
+            }
+            list.tail = pkt;
+        } else {
+            self.inner = Some(MsgBlkChainInner { head: pkt, tail: pkt });
+        }
+    }
+
+    /// Return the head of the underlying `mblk_t` packet chain and
+    /// consume `self`. The caller of this function now owns the
+    /// `mblk_t` segment chain.
+    pub fn unwrap_mblk(mut self) -> Option<NonNull<mblk_t>> {
+        self.inner.take().map(|v| v.head)
+    }
+}
+
+impl Drop for MsgBlkChain {
+    fn drop(&mut self) {
+        // This is a minor variation on MsgBlk's logic. illumos
+        // contains helper functions from STREAMS to just drop a whole
+        // chain.
+        cfg_if! {
+            if #[cfg(all(not(feature = "std"), not(test)))] {
+                // Safety: This is safe as long as the original
+                // `mblk_t` came from a call to `allocb(9F)` (or
+                // similar API).
+                if let Some(list) = &self.inner {
+                    unsafe { ddi::freemsgchain(list.head.as_ptr()) };
+                }
+            } else {
+                while let Some(pkt) = self.pop_front() {
+                    drop(pkt);
+                }
+            }
+        }
+    }
+}
+
+/// An individual illumos `mblk_t` -- a single bytestream
+/// comprised of a linked list of data segments.
+///
+/// To facilitate testing the OPTE core, [`MsgBlk`] is an abstraction for
+/// manipulating network packets in both a `std` and `no_std` environment.
+/// The first is useful for writing tests against the OPTE core engine and
+/// executing them in userland, without the need for standing up a full-blown
+/// virtual machine.
+///
+/// The `no_std` implementation is used when running in-kernel. The
+/// main difference is the `mblk_t` and `dblk_t` structures are coming
+/// from viona (outbound/Tx) and mac (inbound/Rx), and we consume them
+/// via [`Packet::wrap_mblk()`]. In reality this is typically holding
+/// an Ethernet _frame_, but we prefer to use the colloquial
+/// nomenclature of "packet".
+#[derive(Debug)]
+pub struct MsgBlk {
+    pub inner: NonNull<mblk_t>,
+}
+
+impl Deref for MsgBlk {
+    type Target = [u8];
+
+    fn deref(&self) -> &Self::Target {
+        unsafe {
+            let self_ref = self.inner.as_ref();
+            let rptr = self_ref.b_rptr;
+            let len = self_ref.b_wptr.offset_from(rptr) as usize;
+            slice::from_raw_parts(rptr, len)
+        }
+    }
+}
+
+impl DerefMut for MsgBlk {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        unsafe {
+            let self_ref = self.inner.as_mut();
+            let rptr = self_ref.b_rptr;
+            let len = self_ref.b_wptr.offset_from(rptr) as usize;
+            slice::from_raw_parts_mut(rptr, len)
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct MsgBlkNode(mblk_t);
+
+impl Deref for MsgBlkNode {
+    type Target = [u8];
+
+    fn deref(&self) -> &Self::Target {
+        unsafe {
+            let rptr = self.0.b_rptr;
+            let len = self.0.b_wptr.offset_from(rptr) as usize;
+            slice::from_raw_parts(rptr, len)
+        }
+    }
+}
+
+impl DerefMut for MsgBlkNode {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        unsafe {
+            let rptr = self.0.b_rptr;
+            let len = self.0.b_wptr.offset_from(rptr) as usize;
+            slice::from_raw_parts_mut(rptr, len)
+        }
+    }
+}
+
+impl MsgBlkNode {
+    /// Shrink the writable/readable area by shifting the `b_rptr` by
+    /// `len`; effectively removing bytes from the start of the packet.
+    ///
+    /// # Errors
+    ///
+    /// `SegAdjustError::StartPastEnd`: Shifting the read pointer by
+    /// `len` would move `b_rptr` past `b_wptr`.
+    pub fn drop_front_bytes(&mut self, n: usize) -> Result<(), SegAdjustError> {
+        unsafe {
+            if self.0.b_wptr.offset_from(self.0.b_rptr) < n as isize {
+                return Err(SegAdjustError::StartPastEnd);
+            }
+            self.0.b_rptr = self.0.b_rptr.add(n);
+        }
+
+        Ok(())
+    }
+}
+
+impl MsgBlk {
+    /// Allocate a new [`MsgBlk`] containing a data buffer of `len`
+    /// bytes.
+    ///
+    /// The returned packet consists of exactly one segment.
+    ///
+    /// In the kernel environment this uses `allocb(9F)` and
+    /// `freemsg(9F)` under the hood.
+    ///
+    /// In the `std` environment this uses a mock implementation of
+    /// `allocb(9F)` and `freeb(9F)`, which contains enough scaffolding
+    /// to satisfy OPTE's use of the underlying `mblk_t` and `dblk_t`
+    /// structures.
+    pub fn new(len: usize) -> Self {
+        let inner = NonNull::new(allocb(len))
+            .expect("somehow failed to get an mblk...");
+
+        Self { inner }
+    }
+
+    /// Allocates a new [`MsgBlk`] of size `buf.len()`, copying its
+    /// contents.
+    pub fn copy(buf: impl AsRef<[u8]>) -> Self {
+        let mut out = Self::new(buf.as_ref().len());
+        // Unwrap safety -- just allocated length of input buffer.
+        out.write_bytes_back(buf).unwrap();
+        out
+    }
+
+    /// Creates a new [`MsgBlk`] using a given set of packet headers.
+    pub fn new_pkt(emit: impl Emit + EmitDoesNotRelyOnBufContents) -> Self {
+        let mut pkt = Self::new(emit.packet_length());
+        pkt.emit_back(emit).unwrap();
+        pkt
+    }
+
+    /// Returns the number of bytes available for writing before
+    pub fn headroom(&self) -> usize {
+        unsafe {
+            let inner = self.inner.as_ref();
+
+            inner.b_rptr.offset_from((*inner.b_datap).db_base) as usize
+        }
+    }
+
+    /// Creates a new [`MsgBlk`] containing a data buffer of `len`
+    /// bytes with 2B of headroom/alignment.
+    ///
+    /// This sets up 4B alignment on all post-ethernet headers.
+    pub fn new_ethernet(len: usize) -> Self {
+        Self::new_with_headroom(2, len)
+    }
+
+    /// Creates a new [`MsgBlk`] using a given set of packet headers
+    /// with 2B of headroom/alignment.
+    ///
+    /// This sets up 4B alignment on all post-ethernet headers.
+    pub fn new_ethernet_pkt(
+        emit: impl Emit + EmitDoesNotRelyOnBufContents,
+    ) -> Self {
+        let mut pkt = Self::new_ethernet(emit.packet_length());
+        pkt.emit_back(emit).unwrap();
+        pkt
+    }
+
+    /// Return the number of initialised bytes in this `MsgBlk` over
+    /// all linked segments.
+    pub fn byte_len(&self) -> usize {
+        self.iter().map(|el| el.len()).sum()
+    }
+
+    /// Return the number of initialised bytes in this `MsgBlk` in
+    /// the head segment.
+    pub fn seg_len(&self) -> usize {
+        self.iter().count()
+    }
+
+    /// Allocate a new [`MsgBlk`] containing a data buffer of size
+    /// `head_len + body_len`.
+    ///
+    /// The read/write pointer is set to have `head_len` bytes of
+    /// headroom and `body_len` bytes of capacity at the back.
+    pub fn new_with_headroom(head_len: usize, body_len: usize) -> Self {
+        let mut out = Self::new(head_len + body_len);
+
+        // SAFETY: alloc is contiguous and always larger than head_len.
+        let mut_out = unsafe { out.inner.as_mut() };
+        mut_out.b_rptr = unsafe { mut_out.b_rptr.add(head_len) };
+        mut_out.b_wptr = mut_out.b_rptr;
+
+        out
+    }
+
+    /// Provides a slice of length `n_bytes` at the back of an [`MsgBlk`]
+    /// (if capacity exists) to be initialised, before increasing `len`
+    /// by `n_bytes`.
+    ///
+    /// # Safety
+    /// Users must write a value to every element of the `MaybeUninit`
+    /// buffer at least once in the `MsgBlk` lifecycle -- all `n_bytes`
+    /// are assumed to be initialised.
+    pub unsafe fn write_back(
+        &mut self,
+        n_bytes: usize,
+        f: impl FnOnce(&mut [MaybeUninit<u8>]),
+    ) -> Result<(), WriteError> {
+        let mut_out = unsafe { self.inner.as_mut() };
+        let avail_bytes =
+            unsafe { (*mut_out.b_datap).db_lim.offset_from(mut_out.b_wptr) };
+
+        if avail_bytes < 0 || (avail_bytes as usize) < n_bytes {
+            return Err(WriteError::NotEnoughBytes {
+                available: avail_bytes.max(0) as usize,
+                needed: n_bytes,
+            });
+        }
+
+        let in_slice = unsafe {
+            slice::from_raw_parts_mut(
+                mut_out.b_wptr as *mut MaybeUninit<u8>,
+                n_bytes,
+            )
+        };
+
+        f(in_slice);
+
+        mut_out.b_wptr = unsafe { mut_out.b_wptr.add(n_bytes) };
+
+        Ok(())
+    }
+
+    /// Provides a slice of length `n_bytes` at the front of an [`MsgBlk`]
+    /// (if capacity exists) to be initialised, before increasing `len`
+    /// by `n_bytes`.
+    ///
+    /// # Safety
+    /// Users must write a value to every element of the `MaybeUninit`
+    /// buffer at least once in the `MsgBlk` lifecycle -- all `n_bytes`
+    /// are assumed to be initialised.
+    pub unsafe fn write_front(
+        &mut self,
+        n_bytes: usize,
+        f: impl FnOnce(&mut [MaybeUninit<u8>]),
+    ) -> Result<(), WriteError> {
+        let mut_out = unsafe { self.inner.as_mut() };
+        let avail_bytes =
+            unsafe { mut_out.b_rptr.offset_from((*mut_out.b_datap).db_base) };
+
+        if avail_bytes < 0 || (avail_bytes as usize) < n_bytes {
+            return Err(WriteError::NotEnoughBytes {
+                available: avail_bytes.max(0) as usize,
+                needed: n_bytes,
+            });
+        }
+
+        let new_head = unsafe { mut_out.b_rptr.sub(n_bytes) };
+
+        let in_slice = unsafe {
+            slice::from_raw_parts_mut(new_head as *mut MaybeUninit<u8>, n_bytes)
+        };
+
+        f(in_slice);
+
+        mut_out.b_rptr = new_head;
+
+        Ok(())
+    }
+
+    /// Adjusts the write pointer for this MsgBlk, initialising any extra bytes to 0.
+    pub fn resize(&mut self, new_len: usize) -> Result<(), WriteError> {
+        let len = self.len();
+        if new_len < len {
+            unsafe {
+                let mut_inner = self.inner.as_mut();
+                mut_inner.b_wptr = mut_inner.b_wptr.sub(len - new_len);
+            }
+            Ok(())
+        } else if new_len > len {
+            unsafe {
+                self.write_back(new_len - len, |v| {
+                    // MaybeUninit::fill is unstable.
+                    let n = v.len();
+                    v.as_mut_ptr().write_bytes(0, n);
+                })
+            }
+        } else {
+            Ok(())
+        }
+    }
+
+    /// Emits an `ingot` packet after any bytes present in this mblk.
+    pub fn emit_back(
+        &mut self,
+        pkt: impl Emit + EmitDoesNotRelyOnBufContents,
+    ) -> Result<(), WriteError> {
+        unsafe {
+            self.write_back(pkt.packet_length(), |v| {
+                // Unwrap safety: write will return an Error if
+                // unsuccessful.
+                pkt.emit_uninit(v).unwrap();
+            })
+        }
+    }
+
+    /// Emits an `ingot` packet before any bytes present in this mblk.
+    pub fn emit_front(
+        &mut self,
+        pkt: impl Emit + EmitDoesNotRelyOnBufContents,
+    ) -> Result<(), WriteError> {
+        unsafe {
+            self.write_front(pkt.packet_length(), |v| {
+                pkt.emit_uninit(v).unwrap();
+            })
+        }
+    }
+
+    /// Copies a byte slice into the region after any bytes present in this mblk.
+    pub fn write_bytes_back(
+        &mut self,
+        bytes: impl AsRef<[u8]>,
+    ) -> Result<(), WriteError> {
+        let bytes = bytes.as_ref();
+        unsafe {
+            self.write_back(bytes.len(), |v| {
+                // feat(maybe_uninit_write_slice) -> copy_from_slice
+                // is unstable.
+                let uninit_src: &[MaybeUninit<u8>] =
+                    core::mem::transmute(bytes);
+                v.copy_from_slice(uninit_src);
+            })
+        }
+    }
+
+    /// Copies a byte slice into the region before any bytes present in this mblk.
+    pub fn write_bytes_front(
+        &mut self,
+        bytes: impl AsRef<[u8]>,
+    ) -> Result<(), WriteError> {
+        let bytes = bytes.as_ref();
+        unsafe {
+            self.write_front(bytes.len(), |v| {
+                // feat(maybe_uninit_write_slice) -> copy_from_slice
+                // is unstable.
+                let uninit_src: &[MaybeUninit<u8>] =
+                    core::mem::transmute(bytes);
+                v.copy_from_slice(uninit_src);
+            })
+        }
+    }
+
+    /// Places another `MsgBlk` at the end of this packet's
+    /// b_cont chain.
+    pub fn append(&mut self, other: Self) {
+        // Find the last element in the pkt chain
+        // i.e., whose b_cont is null.
+        let mut curr = self.inner.as_ptr();
+        while unsafe { !(*curr).b_cont.is_null() } {
+            curr = unsafe { (*curr).b_cont };
+        }
+
+        unsafe {
+            (*curr).b_cont = other.unwrap_mblk().as_ptr();
+        }
+    }
+
+    /// Drop all bytes and move the cursor to the very back of the dblk.
+    pub fn pop_all(&mut self) {
+        unsafe {
+            (*self.inner.as_ptr()).b_rptr =
+                (*(*self.inner.as_ptr()).b_datap).db_lim;
+            (*self.inner.as_ptr()).b_wptr =
+                (*(*self.inner.as_ptr()).b_datap).db_lim;
+        }
+    }
+
+    /// Returns a shared cursor over all segments in this `MsgBlk`.
+    pub fn iter(&self) -> MsgBlkIter {
+        MsgBlkIter { curr: Some(self.inner), marker: PhantomData }
+    }
+
+    /// Returns a mutable cursor over all segments in this `MsgBlk`.
+    pub fn iter_mut(&mut self) -> MsgBlkIterMut {
+        MsgBlkIterMut { curr: Some(self.inner), marker: PhantomData }
+    }
+
+    /// Return the pointer address of the underlying mblk_t.
+    ///
+    /// NOTE: This is purely to allow passing the pointer value up to
+    /// DTrace so that the mblk can be inspected (read only) in probe
+    /// context.
+    pub fn mblk_addr(&self) -> uintptr_t {
+        self.inner.as_ptr() as uintptr_t
+    }
+
+    /// Return the head of the underlying `mblk_t` segment chain and
+    /// consume `self`. The caller of this function now owns the
+    /// `mblk_t` segment chain.
+    pub fn unwrap_mblk(self) -> NonNull<mblk_t> {
+        let ptr_out = self.inner;
+        _ = ManuallyDrop::new(self);
+        ptr_out
+    }
+
+    /// Wrap the `mblk_t` packet in a [`MsgBlk`], taking ownership of
+    /// the `mblk_t` packet as a result. An `mblk_t` packet consists
+    /// of one or more `mblk_t` segments chained together via
+    /// `b_cont`. When the [`MsgBlk`] is dropped, the
+    /// underlying `mblk_t` segment chain is freed. If you wish to
+    /// pass on ownership you must call the [`MsgBlk::unwrap_mblk()`]
+    /// function.
+    ///
+    /// # Safety
+    ///
+    /// The `mp` pointer must point to an `mblk_t` allocated by
+    /// `allocb(9F)` or provided by some kernel API which itself used
+    /// one of the DDI/DKI APIs to allocate it.
+    ///
+    /// # Errors
+    ///
+    /// * Return [`WrapError::NullPtr`] is `mp` is `NULL`.
+    /// * Return [`WrapError::Chain`] is `mp->b_next` or `mp->b_next` are set.
+    pub unsafe fn wrap_mblk(ptr: *mut mblk_t) -> Result<Self, WrapError> {
+        let inner = NonNull::new(ptr).ok_or(WrapError::NullPtr)?;
+        let inner_ref = inner.as_ref();
+
+        if inner_ref.b_next.is_null() && inner_ref.b_prev.is_null() {
+            Ok(Self { inner })
+        } else {
+            Err(WrapError::Chain)
+        }
+    }
+
+    /// Copy out all bytes within this mblk and its successors
+    /// to a single contiguous buffer.
+    pub fn copy_all(&self) -> Vec<u8> {
+        let mut out = vec![];
+
+        for node in self.iter() {
+            out.extend_from_slice(node)
+        }
+
+        out
+    }
+
+    /// Drops all empty mblks from the start of this chain where possible
+    /// (i.e., any empty mblk is followed by another mblk).
+    pub fn drop_empty_segments(&mut self) {
+        // We should not be creating message block continuations to zero
+        // sized blocks. This is not a generally expected thing and has
+        // caused NIC hardware to stop working.
+        // Stripping these out where possible is necessary.
+        let mut head = self.inner;
+        let mut neighbour = unsafe { (*head.as_ptr()).b_cont };
+
+        while !neighbour.is_null()
+            && unsafe { (*head.as_ptr()).b_rptr == (*head.as_ptr()).b_wptr }
+        {
+            // Replace head with neighbour.
+            // Disconnect head from neighbour, and drop head.
+            unsafe {
+                (*head.as_ptr()).b_cont = ptr::null_mut();
+                drop(MsgBlk::wrap_mblk(head.as_ptr()));
+
+                // SAFETY: we know neighbour is non_null.
+                head = NonNull::new_unchecked(neighbour);
+                neighbour = (*head.as_ptr()).b_cont
+            }
+        }
+
+        self.inner = head;
+    }
+}
+
+#[derive(Debug)]
+pub struct MsgBlkIter<'a> {
+    curr: Option<NonNull<mblk_t>>,
+    marker: PhantomData<&'a MsgBlk>,
+}
+
+#[derive(Debug)]
+pub struct MsgBlkIterMut<'a> {
+    curr: Option<NonNull<mblk_t>>,
+    marker: PhantomData<&'a mut MsgBlk>,
+}
+
+impl<'a> MsgBlkIterMut<'a> {
+    pub fn next_iter(&self) -> MsgBlkIter {
+        let curr = self
+            .curr
+            .and_then(|ptr| NonNull::new(unsafe { ptr.as_ref() }.b_cont));
+        MsgBlkIter { curr, marker: PhantomData }
+    }
+
+    pub fn next_iter_mut(&mut self) -> MsgBlkIterMut {
+        let curr = self
+            .curr
+            .and_then(|ptr| NonNull::new(unsafe { ptr.as_ref() }.b_cont));
+        MsgBlkIterMut { curr, marker: PhantomData }
+    }
+}
+
+impl<'a> Iterator for MsgBlkIter<'a> {
+    type Item = &'a MsgBlkNode;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if let Some(ptr) = self.curr {
+            self.curr = NonNull::new(unsafe { (*ptr.as_ptr()).b_cont });
+            // SAFETY: MsgBlkNode is identical to mblk_t.
+            unsafe { Some(&*(ptr.as_ptr() as *const MsgBlkNode)) }
+        } else {
+            None
+        }
+    }
+}
+
+impl<'a> Read for MsgBlkIter<'a> {
+    type Chunk = &'a [u8];
+
+    fn next_chunk(&mut self) -> ingot::types::ParseResult<Self::Chunk> {
+        self.next().ok_or(IngotParseErr::TooSmall).map(|v| v.as_ref())
+    }
+}
+
+impl<'a> Iterator for MsgBlkIterMut<'a> {
+    type Item = &'a mut MsgBlkNode;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if let Some(ptr) = self.curr {
+            self.curr = NonNull::new(unsafe { (*ptr.as_ptr()).b_cont });
+            // SAFETY: MsgBlkNode is identical to mblk_t.
+            unsafe { Some(&mut *(ptr.as_ptr() as *mut MsgBlkNode)) }
+        } else {
+            None
+        }
+    }
+}
+
+impl<'a> Read for MsgBlkIterMut<'a> {
+    type Chunk = &'a mut [u8];
+
+    fn next_chunk(&mut self) -> ingot::types::ParseResult<Self::Chunk> {
+        self.next().ok_or(IngotParseErr::TooSmall).map(|v| v.as_mut())
+    }
+}
+
+impl<'a> QueryLen for MsgBlkIterMut<'a> {
+    #[inline]
+    fn len(&self) -> usize {
+        let own_blk_len = self
+            .curr
+            .map(|v| unsafe {
+                let v = v.as_ref();
+                v.b_wptr.offset_from(v.b_rptr) as usize
+            })
+            .unwrap_or_default();
+
+        own_blk_len + self.next_iter().map(|v| v.len()).sum::<usize>()
+    }
+}
+
+/// For the `no_std`/illumos kernel environment, we want the `mblk_t`
+/// drop to occur at the [`Packet`] level, where we can make use of
+/// `freemsg(9F)`.
+impl Drop for MsgBlk {
+    fn drop(&mut self) {
+        // Drop the segment chain if there is one. Consumers of MsgBlk
+        // will never own a packet with no segments.
+        // This guarantees that we only free the segment chain once.
+        cfg_if! {
+            if #[cfg(all(not(feature = "std"), not(test)))] {
+                // Safety: This is safe as long as the original
+                // `mblk_t` came from a call to `allocb(9F)` (or
+                // similar API).
+                unsafe { ddi::freemsg(self.inner.as_ptr()) };
+            } else {
+                mock_freemsg(self.inner.as_ptr());
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod test {
+    fn create_linked_mblks(n: usize) -> Vec<*mut mblk_t> {
+        let mut els = vec![];
+        for _ in 0..n {
+            els.push(allocb(8));
+        }
+
+        // connect the elements in a chain
+        for (lhs, rhs) in els.iter().zip(els[1..].iter()) {
+            unsafe {
+                (**lhs).b_next = *rhs;
+                (**rhs).b_prev = *lhs;
+            }
+        }
+
+        els
+    }
+
+    #[test]
+    fn chain_has_correct_ends() {
+        let els = create_linked_mblks(3);
+
+        let chain = unsafe { MsgBlkChain::new(els[0]) }.unwrap();
+        let chain_inner = chain.inner.as_ref().unwrap();
+        assert_eq!(chain_inner.head.as_ptr(), els[0]);
+        assert_eq!(chain_inner.tail.as_ptr(), els[2]);
+    }
+
+    #[test]
+    fn chain_breaks_links() {
+        let els = create_linked_mblks(3);
+
+        let mut chain = unsafe { MsgBlkChain::new(els[0]) }.unwrap();
+
+        let p0 = chain.pop_front().unwrap();
+        assert_eq!(p0.mblk_addr(), els[0] as uintptr_t);
+        unsafe {
+            assert!((*els[0]).b_prev.is_null());
+            assert!((*els[0]).b_next.is_null());
+        }
+
+        // Chain head/tail ptrs are correct
+        let chain_inner = chain.inner.as_ref().unwrap();
+        assert_eq!(chain_inner.head.as_ptr(), els[1]);
+        assert_eq!(chain_inner.tail.as_ptr(), els[2]);
+        unsafe {
+            assert!((*els[1]).b_prev.is_null());
+            assert!((*els[2]).b_next.is_null());
+        }
+    }
+
+    #[test]
+    fn chain_append_links() {
+        let els = create_linked_mblks(3);
+        let new_el = allocb(8);
+
+        let mut chain = unsafe { MsgBlkChain::new(els[0]) }.unwrap();
+        let pkt = unsafe { Packet::wrap_mblk(new_el) }.unwrap();
+
+        chain.append(pkt);
+
+        // Chain head/tail ptrs are correct
+        let chain_inner = chain.inner.as_ref().unwrap();
+        assert_eq!(chain_inner.head.as_ptr(), els[0]);
+        assert_eq!(chain_inner.tail.as_ptr(), new_el);
+
+        // Last el has been linked to the new pkt, and it has a valid
+        // backward link.
+        unsafe {
+            assert_eq!((*new_el).b_prev, els[2]);
+            assert!((*new_el).b_next.is_null());
+            assert_eq!((*els[2]).b_next, new_el);
+        }
+    }
+
+    #[test]
+    fn chain_drain_complete() {
+        let els = create_linked_mblks(64);
+
+        let mut chain = unsafe { MsgBlkChain::new(els[0]) }.unwrap();
+
+        for i in 0..els.len() {
+            let pkt = chain.pop_front().unwrap();
+            assert_eq!(pkt.mblk_addr(), els[i] as uintptr_t);
+        }
+
+        assert!(chain.pop_front().is_none());
+    }
+}
