+++ lib/opte/src/engine/icmp/v4.rs
@@ -2,14 +2,14 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-// Copyright 2023 Oxide Computer Company
+// Copyright 2024 Oxide Computer Company
 
 //! ICMPv4 headers and processing.
 
 use super::*;
+use crate::ddi::mblk::MsgBlk;
 use crate::engine::ether::Ethernet;
 use crate::engine::ingot_packet::MblkPacketData;
-use crate::engine::ingot_packet::MsgBlk;
 use crate::engine::ip::v4::Ipv4;
 use crate::engine::ip::L3;
 use crate::engine::predicate::Ipv4AddrMatch;
