+++ lib/opte/src/engine/mod.rs
@@ -39,14 +39,14 @@ pub mod udp;
 
 pub mod ingot_packet;
 
+use crate::ddi::mblk::MsgBlk;
 use checksum::Checksum;
 use ingot::tcp::TcpRef;
 use ingot::types::Read;
 use ingot_packet::FullParsed;
-use ingot_packet::MsgBlk;
 use ingot_packet::OpteMeta;
 use ingot_packet::OpteParsed2;
-use ingot_packet::Packet2;
+use ingot_packet::Packet;
 pub use opte_api::Direction;
 use parse::ValidNoEncap;
 use rule::CompiledTransform;
@@ -221,7 +221,7 @@ pub trait NetworkImpl {
     fn handle_pkt<T: Read>(
         &self,
         dir: Direction,
-        pkt: &mut Packet2<FullParsed<T>>,
+        pkt: &mut Packet<FullParsed<T>>,
         uft_in: &FlowTable<UftEntry<InnerFlowId>>,
         uft_out: &FlowTable<UftEntry<InnerFlowId>>,
     ) -> Result<HdlPktAction, HdlPktError>
